requirements.txt
django==5.2.6
djangorestframework 
psycopg2-binary
python-decouple 
dj-database-url 
PyJWT
markdown
sendgrid
requests
Pillow